# Share environment code between exe and sh backend