… examples/compiled/repl.sh
… tests/_sh/nested_loops.c
… tests/_sh/nested_loops.golden
… tests/_sh/scoping.c
… tests/_sh/scoping.golden