## Context

As we work towards compiling TCC, we'll be making changes to how variables and types are scoped. To make it possible to share those new features between the exe and sh backends, they first need to share their environment code.

Even with the larger code from `env.c`, the `pnut-sh.sh` script is not much larger. This is explained by the removal of the constant parameter optimization (which required a code traversal and by certain code refactors that were allowed by the more powerful environment tracking.

Now:
```
########## pnut-sh ##########
By file:
    3226   10576   81495 pnut.c
    2455   12031   97253 sh.c
     931    4057   37282 sh-runtime.c
     235     755    6578 env.c
    6847   27419  222608 total
By file (without comments or blank lines):
    2433    7639   63669 build/measure/pnut.c
    1829    7880   71874 build/measure/sh.c
     871    3835   35727 build/measure/sh-runtime.c
     196     638    5865 build/measure/env.c
    5329   19992  177135 total
Expanded includes:
    6847   27413  222554 build/measure/pnut-sh.c
    7268   31450  232551 build/measure/pnut-sh.sh
Ratio (Original): 7268/6847 = 1.061
Ratio (Cleaned):  7268/5329 = 1.363
```

Before:
```
########## pnut-sh ##########
By file:
    3226   10576   81495 pnut.c
    2627   12916  104160 sh.c
     931    4057   37282 sh-runtime.c
    6784   27549  222937 total
By file (without comments or blank lines):
    2433    7639   63669 build/measure/pnut.c
    2007    8781   79390 build/measure/sh.c
     871    3835   35727 build/measure/sh-runtime.c
    5311   20255  178786 total
Expanded includes:
    6784   27545  222899 build/measure/pnut-sh.c
    7214   31382  232276 build/measure/pnut-sh.sh
Ratio (Original): 7214/6784 = 1.063
Ratio (Cleaned):  7214/5311 = 1.358
```